fix: separated styles in styles.jsx
