+++ src/components/marginals/Footer/images/index.js
