+++ src/components/marginals/Footer/images/tilted_pend.png
